LayoutTests/media/track/track-webvtt-tc022-entities-expected.txt

Tests special characters &, <, and > in the cue text.


*** Testing text track 0
EXPECTED (cues.length == '3') OK
EXPECTED (cues[i].getCueAsHTML().textContent == 'This cue has an amp & character.') OK
EXPECTED (cues[i].getCueAsHTML().textContent == 'This cue has a less than < character.') OK
EXPECTED (cues[i].getCueAsHTML().textContent == 'This cue has a greater than > character.') OK

*** Testing text track 1
EXPECTED (cues.length == '3') OK
EXPECTED (cues[i].getCueAsHTML().textContent == 'This cue has an amp character. Ampersand is ignored.') OK
EXPECTED (cues[i].getCueAsHTML().textContent == 'This cue has a less than ') OK
EXPECTED (cues[i].getCueAsHTML().textContent == 'This cue has a greater than > character. Since it's not related to a < character, it's just interpreted as text.') OK
END OF TEST
